return integer instead o float

+++ partfinder/analysis_method.py
@@ -176,7 +176,7 @@ class GreedyAnalysis(analysis.Analysis):
                 # this is a fake distance matrix, so that the greedy algorithm
                 # can use all the tricks of the relaxed clustering algorithm
                 dim = len(subsets)
-                d_matrix = np.zeros((((dim * dim) - dim)) / 2)
+                d_matrix = np.zeros(int((((dim * dim) - dim)) / 2))
                 d_matrix[:] = np.inf
 
                 if step == 1:
